Reading magneto meter works somewhat?

+++ DA14531-DA14585-586_Measuring_Temperature/src/temp.c
@@ -1,3 +1,3 @@
 #include "temp.h"
 
-unsigned int sensor_data = 69;
+float sensor_data = 69;
