+++ DA14531-DA14585-586_Measuring_Temperature/src/temp.h
@@ -1,6 +1,6 @@
 #ifndef _TEMP_
 #define _TEMP_
 
-extern unsigned int sensor_data;
+extern float sensor_data;
 
 #endif
